Add labels to the axis labels

A wrapper function for mtext with two changes.

1. The default for text and at is now NULL. If both are NULL, axTicks will be used to determine what values to include on the x (side = 1) or y (side = 2) axis. 2. If length(text) == 1 and at = NULL, then the appropriate side will find the central point of the axis to put the text (e.g., for an axis title).

axis_text(
  text = NULL,
  labels = FALSE,
  side = 3,
  line = 0,
  outer = FALSE,
  at = NULL,
  adj = NA,
  padj = NA,
  cex = NA,
  col = NA,
  font = NA,
  ...
)

Arguments

Examples

if (FALSE) {
   blank(
     xlim = c(0,50),
     ylim = c(0,100),
     bty ="l"
   )
   
   axis_blank(1, at = seq(0,50,10))
   axis_text(side = 1)
   axis_blank(2, at = seq(0,100,20))
   axis_text(side = 2)
   
}