Draw Violins

Draw violins on plots. Determines the values of the variable to evaluate (x_values) and the resulting evaluation values (y_values) of the distribution. these are the raw values

violin(
  x,
  location = NULL,
  rotate = TRUE,
  type = NULL,
  wex = 1,
  values = NULL,
  nvalues = NULL,
  side = "both",
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  ...
)

draw_probs(x, dist_values, violin_values, probs)

draw_violin(values, type = NULL, ...)

violin_values(
  dist_values,
  location = NULL,
  rotate = TRUE,
  type = "l",
  wex = 1,
  side = "both"
)

violin_length_values(dist_values, type = "l", side = "both")

violin_width_values(dist_values, type = "l", wex = 1, side = "both")

dist_values(x, values = NULL, nvalues = NULL, type = "l")

dist_x_values(x, nvalues)

dist_y_values(x, xvals)

default_nvalues(x, type = "l", nrf = NULL, minn = NULL, maxn = NULL)

default_violin_type(x = NULL)

violin_location(location = NULL, rotate = TRUE)

Arguments

Value

violin_location: data.frame of x and y to locate the violin dist_values: data.frame of x and y to from the distribution violin_values: data.frame of x and y to plot